# More MsgPack optimizations