--- a/Fable.Remoting.MsgPack/Fable.Remoting.MsgPack.fsproj
+++ b/Fable.Remoting.MsgPack/Fable.Remoting.MsgPack.fsproj
@@ -13,6 +13,9 @@
     <PackageReleaseNotes>Implement MsgPack serialization in Fable</PackageReleaseNotes>
     <TargetFrameworks>netstandard2.0;net45;net462;netcoreapp3.1</TargetFrameworks>
   </PropertyGroup>
+  <PropertyGroup Condition="'$(TargetFramework)' == 'netcoreapp3.1' ">
+    <DefineConstants>NET_CORE</DefineConstants>
+  </PropertyGroup>
   <ItemGroup>
     <Compile Include="Format.fs" />
     <Compile Include="Write.fs" />