Just playing around with some cool APIs, which I don't get to do at work :))

|                           Type |  Method |        Job | UnrollFactor |         Mean |     Error |    StdDev |       Median |      Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|------------------------------- |-------- |----------- |------------- |-------------:|----------:|----------:|-------------:|-----------:|-----------:|----------:|----------:|
|        Int64ArraySerializationBefore | MsgPack | DefaultJob |           16 |    15.21 ms |  0.112 ms | 0.104 ms |    15.22 ms |   359.3750 |    62.5000 |   62.5000 |   4.29 MB |
|      Int64ArraySerializationAfter |     MsgPack | DefaultJob|16|  11.50 ms | 0.226 ms | 0.211 ms |  11.50 ms|359.3750 |  62.5000 |  62.5000 |   4.29 MB |
|       IntMaybeMapSerializationBefore | MsgPack | DefaultJob |           16 |    11.973 ms | 0.0418 ms | 0.0371 ms |    11.959 ms |   859.3750 |    46.8750 |         - |   7.29 MB |
|       IntMaybeMapSerializationAfter | MsgPack | DefaultJob |           16 |     6.750 ms | 0.0469 ms | 0.0438 ms |     6.756 ms |   671.8750 |    23.4375 |    7.8125 |   5.69 MB |
|   RecursiveRecordSerializationBefore | MsgPack | DefaultJob |           16 |   399.717 ms | 1.3956 ms | 1.3055 ms |   399.781 ms | 14000.0000 |          - |         - |  128.5 MB |
| RecursiveRecordSerializationAfter |     MsgPack |DefaultJob|16| 142.02 ms | 0.676 ms | 0.632 ms |-| 6000.0000 |        - |        - |  64.43 MB |