@Zaid-Ajaj 2020-07-19
These numbers are amazing @kerams 🤯 I would lie to you if I say that I really understand what is going on there. Though I see that the tests added are only running within a dotnet host with Expecto and not with Node and Fable.Mocha correct? If possible, it would be great to also add some integration tests with a couple of test samples like the ones you are adding for `MsgPack.Tests`. Ignore me of course if you already covered those cases with the existing integration tests and I didn't notice :smile: