@kerams 2020-07-19
I believe many tests overlap (no problem with that) and are spread across these files/projects:
- Fable.Remoting.MsgPack.Tests - runs on .NET - checks that serialization -> deserialization produces an object identical to the input
- Fable.Remoting.Giraffe.Tests/MiddlewareTests.fs - runs on .NET - integration test for both Giraffe and .NET client. Whatever test works in MsgPack.Test is pretty much guaranteed to work here too (because both ends are .NET again)
- ClientV2/App.fs - runs with Mocha
  - test list `binaryServerTests` - integration tests for the JS client and Suave server (UITests.fsproj). What's serialized in Suave has to be correctly deserialized in JS
  - test list `msgPackTests` - the JS only counterpart to MsgPack.Tests, important because of major differences in (de)serialization code between .NET and JS (conditional compilation)

Apart from fixing some units of measure serialization issues (tests for which have been added to App.fs), I have only touched the .NET serialization parts in this PR. There are some extra test cases in MsgPack.Tests, but you're right in that it wouldn't hurt to have them covered in `binaryServerTests` too. I'll add it tomorrow.

Feel free to ask if you have specific questions about the implementation.