@kerams 2020-07-20
Note to self: next time run the tests you have just written before pushing to save yourself the embarrassment.

Had to remove 2 tests though:
- Request with `Some None` is serialized to JSON as `null`, which `tryCreateArgsFromJson` correctly parses as just `None`
- Request with `ValueSome "yupp"` is serialized to JSON as `"yupp"`, but `tryCreateArgsFromJson` fails