@Zaid-Ajaj 2020-07-21
> Feel free to ask if you have specific questions about the implementation

Thanks a lot :heart: I will go and learn the MsgPack specs first, that will makes things easier to understand 

> Had to remove 2 tests though:

Yeah I will make separate issues out of those. Like you said, probably no one uses either but it is good to have them tracked for the sake of completeness :100: