@Zaid-Ajaj 2020-07-21
Published a batch of all the libraries (except for Fable.Remoting.Json ofc) which include these optimizations :smile: